fix(calling): improve video state update in a call (#867)
* fix(calling): video state of a participant is being updated in two different places
